app/serializers.py
from rest_framework import serializers
from .models import Dataset

class DatasetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Dataset
        fields = ['id', 'name', 'uploaded_at', 'row_count', 'summary', 'sample_rows']
