requirements.txt
Django>=4.2
djangorestframework
pandas
reportlab
matplotlib
python-multipart
gunicorn
django-cors-headers
